refactor: `Lottos.concat()` 기능 추가
