+++ src/main/java/domain/LottoGenerator.java
@@ -2,6 +2,8 @@ package domain;
 
 import java.util.List;
 
+import domain.lotto.Lotto;
+
 public interface LottoGenerator {
 	List<Lotto> generateLottos(int lottoCount);
 }
