+++ src/main/java/domain/lotto/LottoNumber.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lotto;
 
 import java.util.Objects;
