+++ src/main/java/domain/lotto/Lottos.java
@@ -1,0 +1,33 @@
+package domain.lotto;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import domain.WinningResult;
+
+public class Lottos {
+	private final List<Lotto> lottos;
+
+	public Lottos(List<Lotto> lottos) {
+		this.lottos = lottos;
+	}
+
+	public WinningResult calculateWinningResult(WinningLotto winningLotto) {
+		WinningResult result = new WinningResult();
+		getLottos().forEach(lotto -> result.add(winningLotto.getRank(lotto)));
+		return result;
+	}
+	
+	public Lottos concat(Lottos lottos) {
+		return new Lottos(Stream.concat(this.lottos.stream(),
+				lottos.getLottos().stream())
+			.collect(Collectors.toList()));
+	}
+	
+	public List<Lotto> getLottos() {
+		return Collections.unmodifiableList(lottos);
+	}
+}
