+++ src/test/java/domain/LottoGeneratorTest.java
@@ -9,6 +9,9 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import domain.lotto.Lotto;
+import domain.lotto.LottoNumber;
+
 public class LottoGeneratorTest {
 	@Test
 	@DisplayName("로또생성기는 1부터 45까지, 6개의 서로다른 숫자로 이뤄진 로또 번호를 생성한다.")
