+++ src/test/java/domain/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lotto;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,6 +8,9 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import domain.lotto.Lotto;
+import domain.lotto.LottoNumber;
 
 public class LottoTest {
